+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to the "al-codeactions" extension will be documented in this file.
 
+## 0.1.3
+
+- Allow a semicolon at the end of a procedure declaration line (e.g. `trigger OnOpenPage();`)
+- Create procedures locally, if possible
+
+## 0.1.2
+
+- Do not create a procedure for an enum-object
+- fix problem with calling procedures that contain parameters with closing brackets (for example an Option-String)
+
 ## 0.1.0
 
 - Preview release